% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_species.R
\name{aggregate_by_species}
\alias{aggregate_by_species}
\title{Aggregate spectra by species}
\usage{
aggregate_by_species(spec)
}
\arguments{
\item{spec}{Spectra matrix where the first column is labeled "species" and
holds species names, and the remaining columns are wavelengths.}
}
\value{
List of 3 matrices: "mean", "sd" for standard deviation and "se" for standard error
}
\description{
Aggregate spectra by species
}

