% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spec_evol.R
\name{fit_spec_evol}
\alias{fit_spec_evol}
\title{Fit evolutionary model to spectra}
\usage{
fit_spec_evol(tree, spec, model = "BM", ...)
}
\arguments{
\item{tree}{A phylogeny of class "phylo"}

\item{spec}{A spectra dataset, where the first column are species names and
the remaining are data.}

\item{model}{One of the following models: "BM" or "OU"}
}
\description{
Fit evolutionary model to spectra
}

