% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_spec_on_tree.R
\name{i_add_tip_everywhere}
\alias{i_add_tip_everywhere}
\title{Add a tip to every branch on a phylogeny}
\usage{
i_add_tip_everywhere(base_tree, tip_label)
}
\arguments{
\item{base_tree}{Tree of class "phylo" where the new tip will be grafted.}

\item{tip_label}{Name of the new tip. Must be different from the tip labels
in `base_tree`}
}
\description{
Add a tip to every branch on a phylogeny
}

