% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_spec_on_tree.R
\name{i_place_spec_on_tree_ml}
\alias{i_place_spec_on_tree_ml}
\title{Place specrum on tree using maximum likelihood}
\usage{
i_place_spec_on_tree_ml(tree, spec, model = "BM")
}
\arguments{
\item{tree}{Tree of class "phylo".}

\item{spec}{Spectral dataset, where the first column are species names and
the remaining are data. The spectra with unknown placement must be in this
data.}

\item{model}{Choose "BM" or ...}
}
\description{
Place specrum on tree using maximum likelihood
}

