% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oak_spec_jcb2016}
\alias{oak_spec_jcb2016}
\title{Oaks of North America spectral data}
\format{Matrix where the forst column are species names and the remaining 2001 columns are
wavelengths 400 to 2400 nm. Spectral resolution = 1nm.}
\source{
Jeannine Cavender-Bares. **todo**{link to paper and dryad}
}
\usage{
oak_spec_jcb2016
}
\description{
ASD spectral measurements of Oaks grown in a greenhouse at UMN.
}
\keyword{datasets}

