% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oak_tree}
\alias{oak_tree}
\title{Phylogeny for Oaks of North America}
\format{An object of class "phylo" ("ape" package).}
\source{
Hipp et al.
}
\usage{
oak_tree
}
\description{
Part of the Oak time calibrated tree from Hipp et al. (unpublished).
}
\keyword{datasets}

