% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_signal.R
\name{phylo_signal}
\alias{phylo_signal}
\title{Estimates phylogenetic signal on spectra}
\usage{
phylo_signal(tree, spec, method = c("lambda", "K"), by_band = TRUE,
  se = NULL)
}
\arguments{
\item{tree}{A phylogeny of class "phylo"}

\item{spec}{A matrix with spectral data, where the first column is are species
names that match tip names of the tree.}

\item{method}{Either "lambda" (Pagel's) or "K" (Bloomberg's).}

\item{by_band}{Boolean. The default TRUE estimates signal for each wavelength
independently. If FALSE, signal is estimated for the spectrum as a whole using
method = "lambda" ("K" not availiable if by_band = FALSE).}

\item{se}{Optional. A matrix of spectral standard erros returned by
aggregate_by_species`}
}
\value{
A two column data.frame with the estimated signal in the first column
and a p-value from a significance test in the second column.
}
\description{
Estimates phylogenetic signal on spectra
}

