% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_spec_on_tree.R
\name{place_spec_on_tree}
\alias{place_spec_on_tree}
\title{Places an ulabebed specrum on a tree}
\usage{
place_spec_on_tree(tree, spec, model = "BM", method = "ml")
}
\arguments{
\item{tree}{Tree of class "phylo".}

\item{spec}{pectral dataset, where the first column are species names and
the remaining are data. The spectra with unknown placement must be in this}

\item{model}{Only "BM" availiable for now.}

\item{method}{Choose "ml" or "bayes"}
}
\description{
Places an ulabebed specrum on a tree
}

