% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_autocov_estimate}
\alias{spec_autocov_estimate}
\title{Covariance between all spectral bands}
\format{Matrix of dimension 2001 by 2001. Covariance between bands 400 to 2400
at 1nm resolution.}
\source{
Phil Townsend \url{http://labs.russell.wisc.edu/townsend/phil-townsend/}
}
\usage{
spec_autocov_estimate
}
\description{
Holds the estimated full symmetric covariance matrix of spectral bands from
an ASD dataset including broadleaf and needle leaf species.
}
\keyword{datasets}

